Move data types to data folder
