+++ src/components/Organizers.astro
@@ -1,5 +1,5 @@
 ---
-import type { Links } from '@/utils/data'
+import type { Links } from '@/data/types'
 import SocialLinks from './SocialLinks.astro'
 
 interface Organizer {
