+++ src/components/Sponsor.astro
@@ -1,5 +1,5 @@
 ---
-import type { Sponsor } from '@/utils/data'
+import type { Sponsor } from '@/data/types'
 
 const props = Astro.props as { sponsor?: Sponsor } & Partial<Sponsor>
